Adding acknowledgments section to all three profiles
Extracted a list of all editors of earlier working documents and listed those individuals in an acknowledgements section at the top of the cloud, web, and mobile specification documents. Note that this is an incomplete list of all participants in the working sessions so some contributors may unfortunately be omitted.